assets: add OAuth2 authentication and authorisation
Add OAuth2-based authentication and authorisation to the assets
application. In the fullness of time, a lot of this can be migrated out
to a separate application but for the moment let it live here to bake a
little.

At its core, this commit adds OAuth2TokenAuthentication as an
authentication class to the view. This class will examine the incoming
request for an OAuth2 token, verify it and create a corresponding Django
user if necessary.

After token verification, the HasScopesPermission checks that the token
has the scopes required by the view. We add both a assetregister and
lookup:anonymous scope since we envisage that the backend will have to
talk to the local lookup API to verify CRSids.

Other than adding the classes above, this commit customises the drf-yasg
swagger UI to allow OAuth2 login and adds tests that authorisation is
required on the view.