assets: increate authentication test coverage
The authentication tests did not test that an expired token results in
the OAuth2 session being re-created. Modify the existing mock to mock
the OAuth2Session object itself so that _get_session() is actually
called.

Add a test which mocks the cached session inside _get_session() to raise
a TokenExpiredError which should result in the cached session being
re-created and the token being validated successfully.