HTTP DELETE should make is_deleted = True but not delete an asset
Create deleted_at that fills in when you try to delete the asset. Is a datetime field.
Modify the ModelView delete function to not delete the asset but instead to fill in the deleted_at field.
Modify the list method from the ModelView to not list the assets that have deleted_at not null

Closes #13