respect Django {add,change,delete}_asset permissions
If a user already has the add_asset, change_asset and/or delete_asset
permission, they should be allowed to modify any asset. This allows us
to configure sufficiently privileged users to be able to administer
assets.

Unfortunately this doesn't solve the machine-to-machine case as a
Django user is not created when a token is presented with no subject. We
should find a way to map machine users to Django users.

This change supports having the backend provide some indication to the
frontend about asset permissions since the frontend cannot hope to know
about administrator users.