rework lookup response caching
IMPORTANT: this commit removes the previous LOOKUP_SELF setting and
replaces it with LOOKUP_ROOT which points to the root of the lookupproxy
API server. Deployments will need to be re-configured.

Previously we (ab)used the token passed to us from the user to retrieve
their lookup profile. The token is only conveniently present at
authorisation time and so the authorisation logic was combined with
fetching and caching the lookup resource for the current user.

Aside from complication the authorisation flow, it slightly violated the
spirit of OAuth2 tokens in that we were not accessing lookup *on behalf*
of the user. We were, in fact, accessing lookup on behalf of the IAR
backend application.

We also encoded the lookup scheme and identifier in the username which,
if not exactly brittle, was ugly.

Re-work the lookup caching logic to be separate from the authentication
logic:

1. Explicitly record the mapping between lookup scheme/identifier pairs
   and Django users by way of a new UserLookup model.

2. Separate the caching logic out into a new accessor function,
   assets.lookup.get_person_for_user(), which takes a Django user and
   returns the lookup person resource for that user. The resource is
   fetched from an in-memory cache if present or fetched from lookup if
   not. When fetching data from lookup, the IAR backend obtains its own
   token with a lookup:anonymous scope and uses that token to talk to
   lookup. The avoids the need to keep the user's token around.

None of the existing permissions logic has been changed but now we
explicitly ask for a lookup resource for a user *at the point we need
it* rather than relying on it having been cached as part of the
authentication flow.

Since the user token no longer requires a lookup:anonymous scope, remove
it from the Swagger UI.

The default cached person resource lifetime is configurable through a
setting.

We update the scripts/create-client.sh script to create a backend client
capable of requesting hydra.introspect and lookup:anonymous scopes and a
client used for testing the UI.