reflect permissions in asset resources
Add a new field to the returned resource, "allowed_methods", which
reflects the HTTP methods which are allowed (in principle) for the asset
resource.

GET, HEAD and OPTIONS are not included in the returned list because
their presence is implied by being able to fetch the asset.

The PUT, PATCH and DELETE methods depend on the permissions set on the
AssetViewSet but do not set the request body so PUT being in the
"allowed_methods" field of the asset resource means that in principle a
PUT is allowed but only with the correct resource.

Add lines to the existing tests which checks that the methods appear as
required in the resource returned by GET-ing the asset.