lookup.get_person_for_user: return cached result earlier
We inadvertently returned the cached return value for
get_person_for_user *after* touching the database, not before.

Installing the silk profiler revealed that for each index request we
were make 382 queries to the database when using a copy of database with
~1000 assets. This took around 200ms.

By re-ordering these statements, we have reduced the number of database
queries to 5 which took around 10ms.

This is a 20x speedup.

I suggest we use silk going forward :).